In-memory caching for event synchronisation
Currently we hit ESI with each request to calendar's feed endpoint.
ESI is a shared resource and we should optimize for minimum
consumption of resources.

With this commit
* events are synchronised at most once every 90 minutes
* consequent requests to calendar feeds won't hit ESI every time

For caching we use an in-memory store. This means that the cache will
be invalidated with each deploy, dyno restart or process crash. This
should be OK for now.

Closes #130