Handle deleted events when pulling event details
An attempt to pull the details for a non-existing event, results with an
error `EveOnline::Exceptions::ResourceNotFound` i.e the event has been
deleted.

The job will now rescue `EveOnline::Exceptions::ResourceNotFound` errors
and delete the event from the database.

The event is deleted because:

* It keeps it simple when it comes to serving up-to-date event feeds
* We're not in the business of collecting data

There is also a possibility that the character does not exist. However, in
this scenario, errors will be reported by other character-related operations
so we'll keep things simple and not handle this case.