Pull events for new characters after signin
Characters signing for the first time will not see any upcoming events.
This is because we pull events in a scheduled background job.

This creates bad experience for newly signed up users as their calendar
is empty. Additionally, they may think the tool does not work as advertised.

To improve the experience, upcoming events will be pulled right after
new characters sign in.

Closes #167