Fix deployment error
An attempt to deploy results in an error. Heroku attempts to compile the
slug and fails with:

  LoadError: cannot load such file -- standard/rake

In Rakefile, we instruct Rails to load standard's task but the library
is available only during development and test.

The undesired behavior was introduced in 7d1ef0afd3e35f659570b946a9b5c0e33ebcf3cb.